+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## next
+- Add `icy_sixel` feature that uses a Rust implementation of Sixel
 - Remove `lazy_static` dependency in favor of `std::sync::LazyLock`
 - MSRV is now 1.80
+- Use sixel if found in device attributes instead of static TERM list
 - Dont Error in kitty if the temporary file has been deleted by the terminal. (Now `KittySupport::Local` is possible again)
 - Properly check for kitty support on terminals via protocol query, instead of static `TERM` environment variable checking.
   - This for example allows new terminals like (KDE)`Konsole`